+++ src/repositories/index.ts
@@ -1,11 +1,12 @@
-import { foundationPreviewCurriculum } from '../data/curriculum'
+import { mvpCurriculum } from '../data/curriculum'
 import { reactDerivedStateExercise } from '../data/exercises/reactDerivedState'
+import { starterExercisePack } from '../data/exercises/starterExercisePack'
 import { LocalExerciseRepository } from './LocalExerciseRepository'
 import { LocalStorageLearnerStateRepository } from './LocalStorageLearnerStateRepository'
 
 export const exerciseRepository = new LocalExerciseRepository(
-  [reactDerivedStateExercise],
-  foundationPreviewCurriculum,
+  [reactDerivedStateExercise, ...starterExercisePack],
+  mvpCurriculum,
 )
 
 export const learnerStateRepository = new LocalStorageLearnerStateRepository()